pyproject.toml
[project]
name = "mcp-twelvedata"
version = "0.1.0"
description = "MCP server for financial analysis using TwelveData API"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "fastmcp>=0.4.1",
    "twelvedata>=1.2.14",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">=3.12"

[project.scripts]
mcp-twelvedata = "mcp_twelvedata.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
# Target Python 3.12
target-version = "py312"
line-length = 88

[tool.ruff.lint]
# Enable Pyflakes (F), Pycodestyle (E, W), Isort (I), and many more
select = ["F", "E", "W", "I", "N", "UP", "B", "A", "C4", "SIM", "ARG", "PT"]
ignore = []

# Allow autofix for all enabled rules (when supported)
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.0.0",
]
